The initial version of the paper @vjovanov
I uploaded the text I have for now. I know it steel needs a lot of work, and that I will have to change many parts of it, but it is some starting point for me. I will work on this further,progressively, and correct the mistakes, but I believe it is better that you can see what am I doing - maybe you candivert me if I choose a wrong direction.